renamed directory of media service to media instead of videoupload
